+++ azure-mgmt-batch/azure/mgmt/batch/models/__init__.py
@@ -18,10 +18,8 @@ from .batch_account_update_parameters import BatchAccountUpdateParameters
 from .batch_account_regenerate_key_parameters import BatchAccountRegenerateKeyParameters
 from .batch_account_keys import BatchAccountKeys
 from .activate_application_package_parameters import ActivateApplicationPackageParameters
-from .application_create_parameters import ApplicationCreateParameters
-from .application_package import ApplicationPackage
 from .application import Application
-from .application_update_parameters import ApplicationUpdateParameters
+from .application_package import ApplicationPackage
 from .batch_location_quota import BatchLocationQuota
 from .resource import Resource
 from .proxy_resource import ProxyResource
@@ -31,9 +29,10 @@ from .certificate import Certificate
 from .certificate_create_or_update_parameters import CertificateCreateOrUpdateParameters
 from .cloud_service_configuration import CloudServiceConfiguration
 from .image_reference import ImageReference
-from .os_disk import OSDisk
 from .windows_configuration import WindowsConfiguration
 from .data_disk import DataDisk
+from .container_registry import ContainerRegistry
+from .container_configuration import ContainerConfiguration
 from .virtual_machine_configuration import VirtualMachineConfiguration
 from .deployment_configuration import DeploymentConfiguration
 from .fixed_scale_settings import FixedScaleSettings
@@ -47,12 +46,14 @@ from .pool_endpoint_configuration import PoolEndpointConfiguration
 from .network_configuration import NetworkConfiguration
 from .task_scheduling_policy import TaskSchedulingPolicy
 from .linux_user_configuration import LinuxUserConfiguration
+from .windows_user_configuration import WindowsUserConfiguration
 from .user_account import UserAccount
 from .metadata_item import MetadataItem
 from .resource_file import ResourceFile
 from .environment_setting import EnvironmentSetting
 from .auto_user_specification import AutoUserSpecification
 from .user_identity import UserIdentity
+from .task_container_settings import TaskContainerSettings
 from .start_task import StartTask
 from .certificate_reference import CertificateReference
 from .application_package_reference import ApplicationPackageReference
@@ -64,6 +65,7 @@ from .operation import Operation
 from .check_name_availability_parameters import CheckNameAvailabilityParameters
 from .check_name_availability_result import CheckNameAvailabilityResult
 from .batch_account_paged import BatchAccountPaged
+from .application_package_paged import ApplicationPackagePaged
 from .application_paged import ApplicationPaged
 from .operation_paged import OperationPaged
 from .certificate_paged import CertificatePaged
@@ -85,6 +87,7 @@ from .batch_management_client_enums import (
     NetworkSecurityGroupRuleAccess,
     ComputeNodeFillType,
     ElevationLevel,
+    LoginMode,
     AutoUserScope,
     CertificateStoreLocation,
     CertificateVisibility,
@@ -101,10 +104,8 @@ __all__ = [
     'BatchAccountRegenerateKeyParameters',
     'BatchAccountKeys',
     'ActivateApplicationPackageParameters',
-    'ApplicationCreateParameters',
-    'ApplicationPackage',
     'Application',
-    'ApplicationUpdateParameters',
+    'ApplicationPackage',
     'BatchLocationQuota',
     'Resource',
     'ProxyResource',
@@ -114,9 +115,10 @@ __all__ = [
     'CertificateCreateOrUpdateParameters',
     'CloudServiceConfiguration',
     'ImageReference',
-    'OSDisk',
     'WindowsConfiguration',
     'DataDisk',
+    'ContainerRegistry',
+    'ContainerConfiguration',
     'VirtualMachineConfiguration',
     'DeploymentConfiguration',
     'FixedScaleSettings',
@@ -130,12 +132,14 @@ __all__ = [
     'NetworkConfiguration',
     'TaskSchedulingPolicy',
     'LinuxUserConfiguration',
+    'WindowsUserConfiguration',
     'UserAccount',
     'MetadataItem',
     'ResourceFile',
     'EnvironmentSetting',
     'AutoUserSpecification',
     'UserIdentity',
+    'TaskContainerSettings',
     'StartTask',
     'CertificateReference',
     'ApplicationPackageReference',
@@ -147,6 +151,7 @@ __all__ = [
     'CheckNameAvailabilityParameters',
     'CheckNameAvailabilityResult',
     'BatchAccountPaged',
+    'ApplicationPackagePaged',
     'ApplicationPaged',
     'OperationPaged',
     'CertificatePaged',
@@ -167,6 +172,7 @@ __all__ = [
     'NetworkSecurityGroupRuleAccess',
     'ComputeNodeFillType',
     'ElevationLevel',
+    'LoginMode',
     'AutoUserScope',
     'CertificateStoreLocation',
     'CertificateVisibility',