+++ azure-mgmt-batch/azure/mgmt/batch/models/auto_user_specification.py
@@ -16,10 +16,8 @@ class AutoUserSpecification(Model):
     """Specifies the parameters for the auto user that runs a task on the Batch
     service.
 
-    :param scope: The scope for the auto user. pool - specifies that the task
-     runs as the common auto user account which is created on every node in a
-     pool. task - specifies that the service should create a new user for the
-     task. The default value is task. Possible values include: 'Task', 'Pool'
+    :param scope: The scope for the auto user. The default value is task.
+     Possible values include: 'Task', 'Pool'
     :type scope: str or ~azure.mgmt.batch.models.AutoUserScope
     :param elevation_level: The elevation level of the auto user. nonAdmin -
      The auto user is a standard user without elevated access. admin - The auto