+++ azure-mgmt-batch/azure/mgmt/batch/models/cloud_service_configuration.py
@@ -24,17 +24,10 @@ class CloudServiceConfiguration(Model):
      Guest OS Releases
      (https://azure.microsoft.com/documentation/articles/cloud-services-guestos-update-matrix/#releases).
     :type os_family: str
-    :param target_os_version: The Azure Guest OS version to be installed on
-     the virtual machines in the pool. The default value is * which specifies
-     the latest operating system version for the specified OS family.
-    :type target_os_version: str
-    :param current_os_version: The Azure Guest OS Version currently installed
-     on the virtual machines in the pool. This may differ from targetOSVersion
-     if the pool state is Upgrading. In this case some virtual machines may be
-     on the targetOSVersion and some may be on the currentOSVersion during the
-     upgrade process. Once all virtual machines have upgraded, currentOSVersion
-     is updated to be the same as targetOSVersion.
-    :type current_os_version: str
+    :param os_version: The Azure Guest OS version to be installed on the
+     virtual machines in the pool. The default value is * which specifies the
+     latest operating system version for the specified OS family.
+    :type os_version: str
     """
 
     _validation = {
@@ -43,12 +36,10 @@ class CloudServiceConfiguration(Model):
 
     _attribute_map = {
         'os_family': {'key': 'osFamily', 'type': 'str'},
-        'target_os_version': {'key': 'targetOSVersion', 'type': 'str'},
-        'current_os_version': {'key': 'currentOSVersion', 'type': 'str'},
+        'os_version': {'key': 'osVersion', 'type': 'str'},
     }
 
-    def __init__(self, os_family, target_os_version=None, current_os_version=None):
+    def __init__(self, os_family, os_version=None):
         super(CloudServiceConfiguration, self).__init__()
         self.os_family = os_family
-        self.target_os_version = target_os_version
-        self.current_os_version = current_os_version
+        self.os_version = os_version