+++ azure-mgmt-batch/azure/mgmt/batch/models/virtual_machine_configuration.py
@@ -19,9 +19,6 @@ class VirtualMachineConfiguration(Model):
     :param image_reference: A reference to the Azure Virtual Machines
      Marketplace Image or the custom Virtual Machine Image to use.
     :type image_reference: ~azure.mgmt.batch.models.ImageReference
-    :param os_disk: Settings for the operating system disk of the Virtual
-     Machine.
-    :type os_disk: ~azure.mgmt.batch.models.OSDisk
     :param node_agent_sku_id: The SKU of the Batch node agent to be
      provisioned on compute nodes in the pool. The Batch node agent is a
      program that runs on each node in the pool, and provides the
@@ -37,7 +34,7 @@ class VirtualMachineConfiguration(Model):
      specifies a Linux OS image.
     :type windows_configuration: ~azure.mgmt.batch.models.WindowsConfiguration
     :param data_disks: The configuration for data disks attached to the
-     comptue nodes in the pool. This property must be specified if the compute
+     compute nodes in the pool. This property must be specified if the compute
      nodes in the pool need to have empty data disks attached to them.
     :type data_disks: list[~azure.mgmt.batch.models.DataDisk]
     :param license_type: The type of on-premises license to be used when
@@ -48,6 +45,13 @@ class VirtualMachineConfiguration(Model):
      Windows_Server - The on-premises license is for Windows Server.
      Windows_Client - The on-premises license is for Windows Client.
     :type license_type: str
+    :param container_configuration: The container configuration for the pool.
+     If specified, setup is performed on each node in the pool to allow tasks
+     to run in containers. All regular tasks and job manager tasks run on this
+     pool must specify the containerSettings property, and all other tasks may
+     specify it.
+    :type container_configuration:
+     ~azure.mgmt.batch.models.ContainerConfiguration
     """
 
     _validation = {
@@ -57,18 +61,18 @@ class VirtualMachineConfiguration(Model):
 
     _attribute_map = {
         'image_reference': {'key': 'imageReference', 'type': 'ImageReference'},
-        'os_disk': {'key': 'osDisk', 'type': 'OSDisk'},
         'node_agent_sku_id': {'key': 'nodeAgentSkuId', 'type': 'str'},
         'windows_configuration': {'key': 'windowsConfiguration', 'type': 'WindowsConfiguration'},
         'data_disks': {'key': 'dataDisks', 'type': '[DataDisk]'},
         'license_type': {'key': 'licenseType', 'type': 'str'},
+        'container_configuration': {'key': 'containerConfiguration', 'type': 'ContainerConfiguration'},
     }
 
-    def __init__(self, image_reference, node_agent_sku_id, os_disk=None, windows_configuration=None, data_disks=None, license_type=None):
+    def __init__(self, image_reference, node_agent_sku_id, windows_configuration=None, data_disks=None, license_type=None, container_configuration=None):
         super(VirtualMachineConfiguration, self).__init__()
         self.image_reference = image_reference
-        self.os_disk = os_disk
         self.node_agent_sku_id = node_agent_sku_id
         self.windows_configuration = windows_configuration
         self.data_disks = data_disks
         self.license_type = license_type
+        self.container_configuration = container_configuration